chore: factor chat logic out to prep for login screen
