+++ react-native/ModeratedChat/utils/user.ts
@@ -1,0 +1,26 @@
+import { type User } from "../shared/models";
+import { v4 } from "uuid";
+
+const usernameLocalStorageKey = "username-v2";
+
+export const getUser = (): User => {
+  return {username: "pete", id: v4()};
+  // const storedUser = localStorage.getItem(usernameLocalStorageKey);
+  // if (!storedUser) {
+  //   throw new Error("User not found");
+  // }
+  // return JSON.parse(storedUser) as User;
+};
+
+export const createUser = (username: string) => {
+  const user: User = {
+    username,
+    id: v4(),
+  };
+  // localStorage.setItem(usernameLocalStorageKey, JSON.stringify(user));
+};
+
+export const doesUserExist = (): boolean => {
+  return true;
+  // return !!localStorage.getItem(usernameLocalStorageKey);
+};
